+++ packages/api/tests/endpoints/granules/test-searchafter-10k.js
@@ -10,7 +10,6 @@ const { loadGranules, granuleFactory } = require('@cumulus/es-client/tests/helpe
 process.env.AccessTokensTable = randomId('token');
 process.env.AsyncOperationsTable = randomId('async');
 process.env.ExecutionsTable = randomId('executions');
-process.env.CollectionsTable = randomId('collection');
 process.env.GranulesTable = randomId('granules');
 process.env.stackName = randomId('stackname');
 process.env.system_bucket = randomId('system-bucket');